64-bit conversion
git-svn-id: svn://orca.physics.unc.edu/Orca64@11898 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Vme/CAEN/CV1721/ORCV1721Model.h
@@ -124,7 +124,7 @@ enum {
     unsigned long   numberBLTEventsToReadout;
 	
 	//cached variables, valid only during running
-	unsigned int    statusReg;
+	unsigned long    statusReg;
 	unsigned long   location;
 	unsigned long	eventSizeReg;
 	unsigned long	dataReg;
